+++ src/devices/HmIPBlind.ts
@@ -64,10 +64,16 @@ export class HmIPBlind extends HmIPShutter implements Updateable {
         if (slatsLevelHomeKit != this.slatsLevel) {
           this.slatsLevel = slatsLevelHomeKit;
           this.platform.log.info('Current blind slats level of %s changed to %s', this.accessory.displayName, this.slatsLevel.toFixed(0));
-          this.service.updateCharacteristic(this.platform.Characteristic.CurrentHorizontalTiltAngle, slatsLevelHomeKit);
+          this.service.updateCharacteristic(this.platform.Characteristic.CurrentHorizontalTiltAngle, this.slatsLevel);
         }
-
       }
+    }
+  }
+
+  protected updateProcessingState() {
+    super.updateProcessingState();
+    if (!this.processing) {
+      this.service.updateCharacteristic(this.platform.Characteristic.TargetHorizontalTiltAngle, this.slatsLevel);
     }
   }